ecom_metrics/requirements.txt
asgiref==3.8.1
black==24.10.0
boto3==1.35.71
botocore==1.35.71
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
